rename package to singular example
