+++ pkg/linguist/data/vendor.yml
@@ -328,7 +328,7 @@
 - (^|/)extjs/docs/
 - (^|/)extjs/builds/
 - (^|/)extjs/cmd/
-- (^|/)extjs/examples/
+- (^|/)extjs/example/
 - (^|/)extjs/locale/
 - (^|/)extjs/packages/
 - (^|/)extjs/plugins/
